Add new feature previews
I have been working on some new features for the program. To recap there
are a few notable ones that you can see but not use. The first being the
Menu Strip. The second being the Numerical Up Down thing. That's gonna
be so you can set your own increment value (so instead of incrementing
by 1 each time, you can increment by 6). As well, if you see in the Menu
Strip, there are place holders for Saving and Loading values. That has
literally no function yet, but they are there when I get to it.

Also if you look at the source code I started working on a way to allow
users to use their own custom input value. I might be missing something
else.

Organized some code as well so it is not too jumbled up.